services/rtsp-service/requirements.txt
# Install all requirement for 
# pre-process step

opencv-python-headless
pymongo
fastapi[standard]
pika
python-dotenv
uvicorn